requirements.txt
django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
python-dotenv
Pillow
# psycopg2-binary  # Uncomment for PostgreSQL
# mysqlclient      # Uncomment for MySQL